# flask_api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Florida Tornado Response AI Analysis API",
    "version": "0.1.0",
    "description": "Hackathon demo API for tornado risk timeline, property risk analysis, AI-style recommendations, notification drafts, and draft work order confirmation."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "Local Flask development server"
    }
  ],
...
  "paths": {
    "/api/risk/timeline": {
      "get": {
        "tags": [
          "Risk"
        ],
        "summary": "Get severe weather timeline",
        "description": "Returns the selected severe weather event timeline. When time is provided, returns the matching timeline point.",
        "responses": {
          "200": {
            "description": "Weather timeline response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "event": {
                      "type": "object"
                    },
                    "timeline": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "meta": {
                      "type": "object"
                    },
                    "timelinePoint": {
                      "type": [
                        "object",
                        "null"
                      ]
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "metadata": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "required": false,
            "description": "Weather event ID. Defaults to the demo forecast hurricane event.",
            "schema": {
              "type": "string",
              "example": "FL-HUR-2026-FCST-01"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "description": "Timeline timestamp to filter by. If omitted, all timeline points for the selected event are returned.",
            "schema": {
              "type": "string",
              "example": "2026-08-17T06:00:00Z"
            }
          }
        ]
      }
    },
    "/api/risk/properties": {
      "get": {
        "tags": [
          "Risk"
        ],
        "summary": "Get affected property risk analysis",
        "description": "Calculates and returns portfolio and property-level risk results for a selected timeline point.",
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "required": false,
            "description": "Optional weather event ID. If omitted, the API selects the event that contains the requested time.",
            "schema": {
              "type": "string",
              "example": "FL-HUR-2026-FCST-01"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "description": "Timeline timestamp. If eventId is omitted, this timestamp is used to select the weather event.",
            "schema": {
              "type": "string",
              "example": "2026-08-17T06:00:00Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Risk analysis response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid timeline point",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/storm-center/properties": {
      "get": {
        "tags": [
          "Risk"
        ],
        "summary": "Get properties near the storm center",
        "description": "Returns properties within radiusKm of the selected storm timeline center point. Defaults radiusKm to the timeline point impactRadiusKm when omitted.",
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "required": false,
            "description": "Optional weather event ID. If omitted, the API selects the event that contains the requested time.",
            "schema": {
              "type": "string",
              "example": "FL-HUR-2026-FCST-01"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "description": "Timeline timestamp used to locate the storm center.",
            "schema": {
              "type": "string",
              "example": "2026-08-17T06:00:00Z"
            }
          },
          {
            "name": "radiusKm",
            "in": "query",
            "required": false,
            "description": "Search radius in kilometers around the storm center. Defaults to the timeline point impactRadiusKm, or 50 when unavailable.",
            "schema": {
              "type": "number",
              "example": 80
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Storm-center related properties response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StormCenterPropertiesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid timeline point or radius",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/recommendations": {
      "post": {
        "tags": [
          "AI Content"
        ],
        "summary": "Generate AI-style recommendations",
        "description": "Returns a property-level recommendation when propertyId is provided; otherwise returns portfolio-level recommendation.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "time": {
                    "type": "string",
                    "example": "2026-08-17T06:00:00Z"
                  },
                  "propertyId": {
                    "type": "string",
                    "example": "FL-TPA-205"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recommendation response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Property not found in affected results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/intelligence": {
      "get": {
        "tags": ["Portfolio Intelligence"],
        "summary": "Read-only Layer 1 Portfolio Intelligence for an analysis scope",
        "description": "Aggregates assetHealthScore, stormImpactLevel, riskScore_v2, lossForecast, insuranceGap, capitalROI (per capital action + bestCapitalAction per property), and the final priorityRanking for the requested analysis scope. Without parameters the default demo scope (FL-DEMO / 2026 / current demo storm) is used. The response includes finalPriorityList (Phase C ranking) alongside the legacy watchList (kept for compatibility).",
        "parameters": [
          {
            "name": "portfolioId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "example": "FL-DEMO",
            "description": "Portfolio to analyze. The demo dataset contains only FL-DEMO; unknown ids are served with a diagnostics warning."
          },
          {
            "name": "analysisYear",
            "in": "query",
            "required": false,
            "schema": { "type": "integer" },
            "example": 2026,
            "description": "Analysis year. Bounds the rolling work-order lookback window and which valuations are considered valid."
          },
          {
            "name": "stormEventId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "example": "TOR-FL-2026-0612",
            "description": "Storm event to analyze. Falls back to the current demo storm (with a diagnostics warning) when the id is unknown."
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped Portfolio Intelligence payload (portfolioSummary, propertyIntelligenceResults, watchList, finalPriorityList, capitalActionResults, diagnostics.analysisScope)"
          },
          "400": {
            "description": "Invalid scope parameter (e.g. non-integer analysisYear)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/ai-copilot/analyze": {
      "post": {
        "tags": ["AI Copilot"],
        "summary": "Run the Layer 3 AI Copilot analysis",
        "description": "Builds AI Copilot State from demo data and deterministic Layer 1/Layer 2 outputs, calls the internal AI Platform Agent when configured, and falls back to a deterministic mock response for local demos.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AiCopilotAnalyzeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI Copilot analysis response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AiCopilotAnalyzeResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid Copilot request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/notifications/draft": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Generate resident notification draft",
        "description": "Returns SMS, push notification, and email drafts for one affected property.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification draft response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDraftResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Property not found in affected results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PropertyRequest": {
        "type": "object",
        "required": [
          "propertyId"
        ],
        "properties": {
          "time": {
            "type": "string",
            "example": "2026-08-17T06:00:00Z"
          },
          "propertyId": {
            "type": "string",
            "example": "FL-TPA-205"
          },
          "eventId": {
            "type": "string",
            "example": "FL-HUR-2026-FCST-01",
            "description": "Optional weather event ID. If omitted, the API selects the event that contains the requested time."
          }
        }
      },
...
      "RiskAnalysisResponse": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "analysisTime": {
            "type": "string"
          },
          "scenarioStage": {
            "type": "string"
          },
          "portfolioSummary": {
            "$ref": "#/components/schemas/PortfolioSummary"
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyRiskResult"
            }
          },
          "llmInputs": {
            "type": "object"
          },
          "dataSources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "StormCenterPropertiesResponse": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "eventName": {
            "type": "string"
          },
          "eventType": {
            "type": "string"
          },
          "analysisTime": {
            "type": "string"
          },
          "scenarioStage": {
            "type": "string"
          },
          "stormCenter": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number"
              },
              "lng": {
                "type": "number"
              },
              "source": {
                "type": "string"
              }
            }
          },
          "radiusKm": {
            "type": "number"
          },
          "matchedProperties": {
            "type": "integer"
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StormCenterPropertyResult"
            }
          },
          "timelinePoint": {
            "type": "object"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
    }
  }
}